// bookspace-client/src/types/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface UserDTO
 */
export interface UserDTO {

    /**
     * @type {string}
     * @memberof UserDTO
     */
    id?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    username?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    email?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    password?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    role?: UserDTORoleEnum;

    /**
     * @type {Date}
     * @memberof UserDTO
     */
    createdAt?: Date;
}

/**
 * @export
 * @enum {string}
 */
export enum UserDTORoleEnum {
    USER = 'USER',
    ADMIN = 'ADMIN'
}